Address review comments
- check map keys in duration constructor
- clean up SparkSQLExpressionMapper
