+++ okapi-api/src/main/scala/org/opencypher/okapi/impl/temporal/Duration.scala
@@ -1,29 +1,29 @@
 /**
- * Copyright (c) 2016-2019 "Neo4j Sweden, AB" [https://neo4j.com]
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Attribution Notice under the terms of the Apache License 2.0
- *
- * This work was created by the collective efforts of the openCypher community.
- * Without limiting the terms of Section 6, any Derivative Work that is not
- * approved by the public consensus process of the openCypher Implementers Group
- * should not be described as “Cypher” (and Cypher® is a registered trademark of
- * Neo4j Inc.) or as "openCypher". Extensions by implementers or prototypes or
- * proposals for change that have been documented or implemented should only be
- * described as "implementation extensions to Cypher" or as "proposed changes to
- * Cypher that are not yet approved by the openCypher community".
- */
+  * Copyright (c) 2016-2019 "Neo4j Sweden, AB" [https://neo4j.com]
+  *
+  * Licensed under the Apache License, Version 2.0 (the "License");
+  * you may not use this file except in compliance with the License.
+  * You may obtain a copy of the License at
+  *
+  * http://www.apache.org/licenses/LICENSE-2.0
+  *
+  * Unless required by applicable law or agreed to in writing, software
+  * distributed under the License is distributed on an "AS IS" BASIS,
+  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  * See the License for the specific language governing permissions and
+  * limitations under the License.
+  *
+  * Attribution Notice under the terms of the Apache License 2.0
+  *
+  * This work was created by the collective efforts of the openCypher community.
+  * Without limiting the terms of Section 6, any Derivative Work that is not
+  * approved by the public consensus process of the openCypher Implementers Group
+  * should not be described as “Cypher” (and Cypher® is a registered trademark of
+  * Neo4j Inc.) or as "openCypher". Extensions by implementers or prototypes or
+  * proposals for change that have been documented or implemented should only be
+  * described as "implementation extensions to Cypher" or as "proposed changes to
+  * Cypher that are not yet approved by the openCypher community".
+  */
 package org.opencypher.okapi.impl.temporal
 
 import java.time.temporal.ChronoUnit
@@ -37,13 +37,13 @@ import org.opencypher.okapi.impl.temporal.TemporalConstants._
 /**
   * Okapi representation of a duration.
   *
-  * @param months number of months
-  * @param days number of days
+  * @param months  number of months
+  * @param days    number of days
   * @param seconds number of seconds
-  * @param nanos normalized number of nanoseconds spanning fractions of a second
+  * @param nanos   normalized number of nanoseconds spanning fractions of a second
   */
-class Duration protected (val months: Long = 0, val days: Long = 0, val seconds: Long = 0, val nanos: Long = 0)
-extends Ordered[Duration] {
+class Duration protected(val months: Long = 0, val days: Long = 0, val seconds: Long = 0, val nanos: Long = 0)
+  extends Ordered[Duration] {
   def toJava: (java.time.Period, java.time.Duration) = {
     val period = java.time.Period.of((months / 12).toInt, (months % 12).toInt, days.toInt)
     val duration = java.time.Duration.ofSeconds(seconds).plus(nanos, ChronoUnit.NANOS)
@@ -54,13 +54,13 @@ extends Ordered[Duration] {
   override def compare(that: Duration): Int = COMPARATOR.compare(this, that)
 
   override def equals(o: Any): Boolean = o match {
-    case d:Duration => compare(d) == 0
+    case d: Duration => compare(d) == 0
     case _ => false
   }
-/*
- * Since not every month has the same amount of seconds, we use the average to sum this duration in seconds.
- * Not every day has the same amount of seconds either, but since there is one day with +1 hour and one with -1 hour the average is still 24*3600 per day.
- */
+  /*
+   * Since not every month has the same amount of seconds, we use the average to sum this duration in seconds.
+   * Not every day has the same amount of seconds either, but since there is one day with +1 hour and one with -1 hour the average is still 24*3600 per day.
+   */
 
   def averageLengthInSeconds: Long = {
     val daysInSeconds = Math.multiplyExact(days, SECONDS_PER_DAY)
@@ -91,9 +91,25 @@ extends Ordered[Duration] {
 }
 
 object Duration {
-  def apply(years: Long = 0, months: Long = 0, weeks: Long = 0, days: Long = 0,
+
+  val SUPPORTED_KEYS = Set(
+    "years",
+    "months",
+    "weeks",
+    "days",
+    "hours",
+    "minutes",
+    "seconds",
+    "milliseconds",
+    "microseconds",
+    "nanoseconds"
+  )
+
+  def apply(
+    years: Long = 0, months: Long = 0, weeks: Long = 0, days: Long = 0,
     hours: Long = 0, minutes: Long = 0,
-    seconds: Long = 0, milliseconds: Long = 0, microseconds: Long = 0, nanoseconds: Long = 0): Duration = {
+    seconds: Long = 0, milliseconds: Long = 0, microseconds: Long = 0, nanoseconds: Long = 0
+  ): Duration = {
 
     val nanoSum = milliseconds * 1000000 + microseconds * 1000 + nanoseconds
     val normalizedSeconds = hours * SECONDS_PER_HOUR + minutes * SECONDS_PER_MINUTE + seconds + nanoSum / NANOS_PER_SECOND
@@ -117,11 +133,20 @@ object Duration {
 
   def apply(map: Map[String, Long]): Duration = {
     val sanitizedMap = map.map { case (key, value) => key.toLowerCase -> value }
-    // TODO: throw error if map key is unsupported
+
+    val unsupportedKeys = sanitizedMap.keySet -- SUPPORTED_KEYS
+    if (unsupportedKeys.nonEmpty) {
+      throw IllegalArgumentException(
+        SUPPORTED_KEYS.mkString(", "),
+        sanitizedMap.keySet.mkString(", "),
+        s"Unsupported duration values: ${unsupportedKeys.mkString(", ")}"
+      )
+    }
+
     Duration(
       years = sanitizedMap.getOrElse("years", 0),
       months = sanitizedMap.getOrElse("months", 0),
-      weeks = sanitizedMap.getOrElse("weeks", 0) ,
+      weeks = sanitizedMap.getOrElse("weeks", 0),
       days = sanitizedMap.getOrElse("days", 0),
       hours = sanitizedMap.getOrElse("hours", 0),
       minutes = sanitizedMap.getOrElse("minutes", 0),
